+++ service/src/main/java/com/dtstack/rdos/engine/service/zk/ZkShardManager.java
@@ -81,7 +81,7 @@ public class ZkShardManager implements Runnable {
                         Iterator<Map.Entry<String, Byte>> it = shardData.entrySet().iterator();
                         while (it.hasNext()) {
                             Map.Entry<String, Byte> data = it.next();
-                            if (RdosTaskStatus.needClean(data.getValue())) {
+                            if (RdosTaskStatus.needClean(data.getValue().intValue())) {
                                 brokerDataShard.getNewVersion().incrementAndGet();
                                 it.remove();
                             }
